telemetry

+++ TelemetryHelper.ps1
@@ -4,10 +4,10 @@
     )
 
     if ($value -is [switch]) {
-        "$($value.IsPresent)"
+        $value.IsPresent
     }
     elseif ($value -is [boolean]) {
-        "$value"
+        $value
     }
     elseif ($value -is [SecureString]) {
         "[SecureString]"
@@ -17,10 +17,10 @@
     }
     elseif ($value -is [string]) {
         if (($value -like "https:*" -or $value -like "http:*") -and ($value.Contains('?'))) {
-            """$($value.Split('?')[0])?[parameters]"""
+            "$($value.Split('?')[0])?[parameters]"
         }
         else {
-            """$value"""
+            "$value"
         }
     }
     elseif ($value -is [System.Collections.IDictionary]) {
@@ -41,7 +41,7 @@
         "$str ]"
     }
     else {
-        "$value"
+        $value
     }
 }
 
@@ -53,12 +53,15 @@ function AddTelemetryProperty {
     )
 
     if ($telemetryScope) {
-#        Write-Host "Telemetry scope $($telemetryScope.Name), add property $key = $((FormatValue -value $value))"
+        $value = FormatValue -value $value
+        if ($telemetry.Debug) {
+            Write-Host -ForegroundColor Yellow "Telemetry scope $($telemetryScope.Name), add property $key = '$value', type = $($value.GetType())"
+        }
         if ($telemetryScope.properties.ContainsKey($Key)) {
-            $telemetryScope.properties."$key" += "`n$(FormatValue -value $value)"
+            $telemetryScope.properties."$key" += "`n$value"
         }
         else {
-            $telemetryScope.properties.Add($key, (FormatValue -value $value))
+            $telemetryScope.properties.Add($key, $value)
         }
     }
 }
@@ -66,9 +69,11 @@ function AddTelemetryProperty {
 function InitTelemetryScope {
     Param(
         [string] $name,
+        [switch] $always,
         [string[]] $includeParameters = @(),
         $parameterValues = $null
     )
+    $includeParameters = "*"
     if ($telemetry.Client) {
         if ($bcContainerHelperConfig.TelemetryConnectionString) {
             if ($telemetry.Client.TelemetryConfiguration.DisableTelemetry -or $telemetry.Client.TelemetryConfiguration.ConnectionString -ne $bcContainerHelperConfig.TelemetryConnectionString) {
@@ -76,26 +81,32 @@ function InitTelemetryScope {
                     try {
                         $telemetry.Client.TelemetryConfiguration.ConnectionString = $bcContainerHelperConfig.TelemetryConnectionString
                         $telemetry.Client.TelemetryConfiguration.DisableTelemetry = $false
-                        Write-Host "Telemetry client initialized"
+                        if ($telemetry.Debug) {
+                            Write-Host -ForegroundColor Yellow "Telemetry client initialized"
+                        }
                     }
                     catch {
                         $telemetry.Client.TelemetryConfiguration.DisableTelemetry = $true
                     }
                 }
             }
-            if ($telemetry.Client.IsEnabled()) {
-                #Write-Host "Init telemetry scope $name"
+            if ($telemetry.Client.IsEnabled() -and ($always -or ($telemetry.CorrelationId -eq ""))) {
+                if ($telemetry.Debug) {
+                    Write-Host -ForegroundColor Yellow "Init telemetry scope $name"
+                }
+                $CorrelationId = [GUID]::NewGuid().ToString()
+                if ($telemetry.TopId -eq "") { $telemetry.TopId = $CorrelationId }
                 $scope = @{
                     "Name" = $name
                     "StartTime" = [DateTime]::Now
                     "SeverityLevel" = 1
                     "Properties" = [Collections.Generic.Dictionary[string, string]]::new()
-                    "CorrelationId" = ""
+                    "CorrelationId" = $CorrelationId
+                    "ParentId" = $telemetry.CorrelationId
+                    "TopId" = $telemetry.TopId
                     "Emitted" = $false
                 }
-                if (!$telemetry.Transcripting) {
-                    $scope.CorrelationId = [GUID]::NewGuid().ToString()
-                }
+                $telemetry.CorrelationId = $CorrelationId
                 if ($includeParameters) {
                     $parameterValues.GetEnumerator() | % {
                         $includeParameter = $false
@@ -109,10 +120,7 @@ function InitTelemetryScope {
                 AddTelemetryProperty -telemetryScope $scope -key "BcContainerHelperVersion" -value $BcContainerHelperVersion
                 AddTelemetryProperty -telemetryScope $scope -key "IsAdministrator" -value $isAdministrator
                 AddTelemetryProperty -telemetryScope $scope -key "StackTrace" -value (Get-PSCallStack | % { "$($_.Command) at $($_.Location)" }) -join "`n"
-                if ($scope.CorrelationId -ne "") {
-                    Start-Transcript -Path (Join-Path $env:TEMP $scope.CorrelationId) | Out-Null
-                    $telemetry.Transcripting = $scope.CorrelationId
-                }
+                Start-Transcript -Path (Join-Path $env:TEMP $CorrelationId) | Out-Null
                 $scope
             }
         }
@@ -128,22 +136,40 @@ function TrackTrace {
     )
 
     if ($telemetryScope -and !$telemetryScope.Emitted) {
-        if ($telemetry.Client.IsEnabled()) {
-            #Write-Host "Emit telemetry trace, scope $($telemetryScope.Name)"
-            $telemetryScope.Properties.Add("Duration", [DateTime]::Now.Subtract($telemetryScope.StartTime).TotalSeconds)
-            $telemetry.Client.TrackTrace($telemetryScope.Name, $telemetryScope.SeverityLevel, $telemetryScope.Properties)
-            $telemetryScope.Emitted = $true
-            if ($telemetry.Transcripting -eq $telemetryScope.CorrelationId) {
-                try{
-                    Stop-Transcript | Out-Null
-                    $telemetry.Transcripting = ""
-                    Remove-Item -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId)
-                }
-                catch {}
+        if ($telemetry.Client.IsEnabled() -and ($telemetryScope.CorrelationId -eq $telemetry.CorrelationId)) {
+            if ($telemetry.Debug) {
+                Write-Host -ForegroundColor Yellow "Emit telemetry trace, scope $($telemetryScope.Name)"
             }
+            $telemetry.CorrelationId = $telemetryScope.ParentId
+            if ($telemetry.CorrelationId -eq "") {
+                $telemetry.TopId = ""
+            }
+            $telemetryScope.Properties.Add("Duration", [DateTime]::Now.Subtract($telemetryScope.StartTime).TotalSeconds)
+            try {
+                Stop-Transcript | Out-Null
+                $transcript = (@(Get-Content -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId)) | select -skip 18 | select -skiplast 4) -join "`n"
+                if ($transcript.Length -gt 30000) {
+                    $transcript = "$($transcript.SubString(0,15000))`n`n...`n`n$($transcript.SubString($transcript.Length-15000))"
+                }
+                Remove-Item -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId)
+            }
+            catch {
+                $transcript = ""
+            }
+
+            $traceTelemetry = $telemetry.Client.GetType().Assembly.CreateInstance('Microsoft.ApplicationInsights.DataContracts.TraceTelemetry')
+            $traceTelemetry.Message = "$($telemetryScope.Name)`n$transcript"
+            $traceTelemetry.SeverityLevel = $telemetryScope.SeverityLevel
+            $telemetryScope.Properties.GetEnumerator() | ForEach-Object { 
+                [void]$traceTelemetry.Properties.TryAdd($_.Key, $_.Value)
+            }
+            $traceTelemetry.Context.Operation.Name = $telemetryScope.Name
+            $traceTelemetry.Context.Operation.Id = $telemetryScope.CorrelationId
+            $traceTelemetry.Context.Operation.ParentId = $telemetryScope.ParentId
+            $telemetry.Client.TrackTrace($traceTelemetry)
+            $telemetryScope.Emitted = $true
         }
     }
-
 }
 
 function TrackException {
@@ -163,27 +189,57 @@ function TrackException {
     )
 
     if ($telemetryScope -and !$telemetryScope.Emitted) {
-        if ($telemetry.Client.IsEnabled()) {
-            #Write-Host "Emit telemetry exception, scope $($telemetryScope.Name)"
+        if ($telemetry.Client.IsEnabled() -and ($telemetryScope.CorrelationId -eq $telemetry.CorrelationId)) {
+            if ($telemetry.Debug) {
+                Write-Host -ForegroundColor Yellow "Emit telemetry exception, scope $($telemetryScope.Name)"
+            }
+            $telemetry.CorrelationId = $telemetryScope.ParentId
+            if ($telemetry.CorrelationId -eq "") {
+                $telemetry.TopId = ""
+            }
             $telemetryScope.Properties.Add("Duration", [DateTime]::Now.Subtract($telemetryScope.StartTime).TotalSeconds)
             if ($scriptStackTrace) {
                 $telemetryScope.Properties.Add("Error StackTrace", $scriptStackTrace)
             }
-   
-            if ($telemetry.Transcripting -eq $telemetryScope.CorrelationId) {
-                try{
-                    Stop-Transcript | Out-Null
-                    $telemetry.Transcripting = ""
-                    $telemetryScope.Properties.Add("Transcript", (Get-Content -Raw -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId)))
-                    $telemetryScope.Properties.Add("CorrelationId", $telemetryScope.CorrelationId)
-                    Write-Host -ForegroundColor Red "$($telemetryScope.Name) failure, Correlation Id: $($telemetryScope.CorrelationId)"
-                    #Get-Content -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId) | Out-Host
-                    Remove-Item -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId)
-    
-                }
-                catch {}
+            if ($exception) {
+                $telemetryScope.Properties.Add("Error Message", $exception.Message)
             }
-            $telemetry.Client.TrackException($exception, $telemetryScope.Properties)
+
+            try {
+                Stop-Transcript | Out-Null
+                $transcript = (@(Get-Content -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId)) | select -skip 18 | select -skiplast 4) -join "`n"
+                if ($transcript.Length -gt 30000) {
+                    $transcript = "$($transcript.SubString(0,15000))`n`n...`n`n$($transcript.SubString($transcript.Length-15000))"
+                }
+                Remove-Item -Path (Join-Path $env:TEMP $telemetryScope.CorrelationId)
+            }
+            catch {
+                $transcript = ""
+            }
+
+            # emit trace telemetry with Error info
+            $traceTelemetry = $telemetry.Client.GetType().Assembly.CreateInstance('Microsoft.ApplicationInsights.DataContracts.TraceTelemetry')
+            $traceTelemetry.Message = "$($telemetryScope.Name)`n$transcript"
+            $traceTelemetry.SeverityLevel = $telemetryScope.SeverityLevel
+            $telemetryScope.Properties.GetEnumerator() | ForEach-Object { 
+                [void]$traceTelemetry.Properties.TryAdd($_.Key, $_.Value)
+            }
+            $traceTelemetry.Context.Operation.Name = $telemetryScope.Name
+            $traceTelemetry.Context.Operation.Id = $telemetryScope.CorrelationId
+            $traceTelemetry.Context.Operation.ParentId = $telemetryScope.ParentId
+            $telemetry.Client.TrackTrace($traceTelemetry)
+
+            # emit exception telemetry
+            $exceptionTelemetry = $telemetry.Client.GetType().Assembly.CreateInstance('Microsoft.ApplicationInsights.DataContracts.ExceptionTelemetry')
+            $exceptionTelemetry.Message = "$($telemetryScope.Name)`n$transcript"
+            $exceptionTelemetry.SeverityLevel = $telemetryScope.SeverityLevel
+            $telemetryScope.Properties.GetEnumerator() | ForEach-Object { 
+                [void]$exceptionTelemetry.Properties.TryAdd($_.Key, $_.Value)
+            }
+            $exceptionTelemetry.Context.Operation.Name = $telemetryScope.Name
+            $exceptionTelemetry.Context.Operation.Id = $telemetryScope.CorrelationId
+            $exceptionTelemetry.Context.Operation.ParentId = $telemetryScope.ParentId
+            $telemetry.Client.TrackException($exceptionTelemetry)
             $telemetryScope.Emitted = $true
         }
     }
